molorag/molorag_standard/requirements.txt
torch
torchvision
transformers
qwen_vl_utils
numpy
pillow
PyMuPDF
networkx
tqdm
scikit-learn
